refactor(multiselect): render via shared select-core with match actions
Multiselect becomes a thin SelectList wrapper. Pinning, freeze-on-open, the
compact-chip trigger, search-scoped bulk rows, and max_selections enforcement
are inherited from the core instead of reimplemented in the plugin.

Also renames components/ui/select/ to components/ui/select-core/ so the new
barrel does not collide with the pre-existing Radix wrapper at select.tsx.
