+++ frontend/src/components/ui/select-core/index.ts
@@ -6,12 +6,7 @@ export {
 } from "./select-list";
 export { useSelectList } from "./use-select-list";
 export { renderSlot, type Slot } from "./render-slot";
-export type {
-  BulkAction,
-  BulkActionSpec,
-  Option,
-  OptionState,
-} from "./types";
+export type { BulkAction, BulkActionSpec, Option, OptionState } from "./types";
 export {
   deselectMatching,
   getBulkActions,
